refactor modules structure and apply clippy
