+++ src/request/body.rs
@@ -1,15 +1,15 @@
 use reqwest::RequestBuilder;
 
+use super::modifier::RequestModifier;
+use super::RequestError;
 use crate::oauth2::OAuth2Config;
 use crate::options::Opts;
-use crate::request::error::RequestError;
-use crate::request::modifier::RequestModifier;
 pub struct Body {}
 
 impl Body {
-  pub fn new()-> Body {
-    Body{}
-  }
+    pub fn new() -> Body {
+        Body {}
+    }
 }
 
 impl RequestModifier for Body {
@@ -19,11 +19,11 @@ impl RequestModifier for Body {
         opts: &Opts,
         _oauth2: &OAuth2Config,
     ) -> Result<RequestBuilder, RequestError> {
-      if let Some(b) = &opts.data {
-        let body = b.clone();
-        Ok(request.body(body))
-      } else {
-        Ok(request)
-      }
+        if let Some(b) = &opts.data {
+            let body = b.clone();
+            Ok(request.body(body))
+        } else {
+            Ok(request)
+        }
     }
 }
