+++ src/request/custom_headers.rs
@@ -1,11 +1,11 @@
 use reqwest::header::{CONTENT_TYPE, USER_AGENT};
 use reqwest::RequestBuilder;
 
+use super::error::RequestError;
+use super::headers::Headers;
+use super::modifier::RequestModifier;
 use crate::oauth2::OAuth2Config;
 use crate::options::Opts;
-use crate::request::error::RequestError;
-use crate::request::headers::Headers;
-use crate::request::modifier::RequestModifier;
 use crate::version;
 
 pub struct CustomHeaders {}
