+++ src/request/mod.rs
@@ -1,11 +1,15 @@
-pub mod access_token_cache;
-pub mod auth_header;
-pub mod body;
-pub mod cors;
-pub mod custom_headers;
-pub mod dispatcher;
-pub mod error;
-pub mod headers;
-pub mod modifier;
-pub mod response;
+mod auth_header;
+mod body;
+mod cors;
+mod custom_headers;
+mod dispatcher;
+mod error;
+mod headers;
+mod modifier;
+mod response;
 mod timeout;
+
+pub use cors::same_origin_redirect_policy;
+pub use dispatcher::Dispatcher;
+pub use error::RequestError;
+pub use response::Response;
